refactored kernels to separate .cc files
